add: HomeHeroコンポーネントをリファクタリングし、ダークモード対応のイントロコンポーネントを追加
